+++ CHANGELOG.md
@@ -123,6 +123,7 @@
 - CONFORMITY CERTIFICATE : new field on printing
 - TIMESHEET : new config to display line numbers
 - Advanced Import : Add support of remove imported records.
+- Business Project : Add 'Project invoicing assistant' batch to update tasks and generate invoicing projects.
 
 ## Improvements
 - Contract: added button to manually close contract if termination date was set in the future